# api/swagger.yaml
swagger: "2.0"
info:
  title: Order MicroService API
  description: Order service APIs for creating and managing orders.
  version: 1.0.0
  contact:
    email: prajjawal.tiwari.0311@gmail.com

basePath: /order

paths:
  /create-order:
    post:
      summary: "Create a new order"
      operationId: "createNewOrder"
      parameters:
        - in: body
          name: body
          description: Order creation details
          required: true
          schema:
            $ref: '#/definitions/CreateOrderRequest'
      responses:
        200:
          description: Order successfully created
          schema:
            $ref: '#/definitions/OrderResponse'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /get-order/{orderId}:
    get:
      summary: "Get order details by order ID"
      operationId: "getOrderDetailsById"
      parameters:
        - name: orderId
          in: path
          description: ID of the order to fetch details
          required: true
          type: string
      responses:
        200:
          description: Order details
          schema:
            $ref: '#/definitions/OrderResponse'
        404:
          description: Order not found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /confirm-order/{orderId}:
    post:
      summary: "Confirm order after payment"
      operationId: "confirmOrderAfterPayment"
      parameters:
        - name: orderId
          in: path
          description: ID of the order to confirm payment
          required: true
          type: string
      responses:
        200:
          description: Order payment confirmed and inventory updated
          schema:
            $ref: '#/definitions/OrderResponse'
        404:
          description: Order not found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Failed to confirm payment or update inventory
          schema:
            $ref: '#/definitions/ErrorModel'

  /send-notification/{orderId}:
    post:
      summary: "Send WhatsApp notification to the owner"
      operationId: "sendNotification"
      parameters:
        - name: orderId
          in: path
          description: ID of the order to send a notification for
          required: true
          type: string
      responses:
        200:
          description: Notification sent successfully
          schema:
            $ref: '#/definitions/OrderResponse'
        404:
          description: Order not found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Failed to send notification
          schema:
            $ref: '#/definitions/ErrorModel'

definitions:
  CreateOrderRequest:
    type: object
    required:
      - customerId
      - products
      - totalAmount
      - address
    properties:
      customerId:
        type: string
      products:
        type: array
        items:
          type: object
          properties:
            productId:
              type: string
            quantity:
              type: integer
      totalAmount:
        type: number
      address:
        type: string

  OrderResponse:
    type: object
    properties:
      customerId:
        type: string
      products:
        type: array
        items:
          type: object
          properties:
            productId:
              type: string
            quantity:
              type: integer
            price:
              type: number
      totalAmount:
        type: number
      address:
        type: string
      status:
        type: string
      paymentConfirmed:
        type: boolean

  ErrorModel:
    type: object
    properties:
      code:
        type: string
      message:
        type: string